Multiple metrics for submission #23 (#38)
* Data structure for multiple metrics for a challenge

We need a new table that combines a test, which is a metric
for a challenge, and a challenge. New table contains results
of the metric calculation.

In addition, there are some cosmetic changes to the tables.

* Cleaning up unused imports and reformatting

* Adjusting challenge creation to new challenge table

* Creating challenge no longer requires ChallengeInput model

* Test table and methods

Methods to store metrics for a challenge in Test table.
There are also methods for adding tests for a challenge and conversion of a json
string into a metric.

* Changed wrong name of submissions table in test model

* Changed wrong type of fields for foreign ids

* Cleaning up unused code

* Tests creation in create challenge endpoint

Previously metrics were kept in the challenges table. Now, each challenge can
have multiple metrics assigned as tests in test table.

* New submission mechanism

When submission is created all tests for a challenge are searched for and for
each an evaluation is done, and saved in evaluation table.

* Cleaning up unused code

* Adding TODOs and removing unused imports

* Fixing new endpoints for frontend to work

* Fixing new endpoints for frontend to work

* Fixing new endpoints for frontend to work

* Fixing new endpoints for frontend to work

* Fixing error in `all_challenges` endpoint while getting score

* Fixing new endpoints for frontend to work

* Fixing new endpoints for frontend to work

* Removing unnecessary imports